// MovieSearchNative/interfaces/Movie.tsx
export interface Movie {
  Series_Title: string;
  IMDB_Rating: string;
  Released_Year: string;
  Certificate: string;
  Director: string;
  Genre: string;
  Gross: string;
  Meta_score: string;
  No_of_Votes: string;
  Overview: string;
  Poster_Link: string;
  Runtime: string;
  Star1: string;
  Star2: string;
  Star3: string;
  Star4: string;
}